services/requirements.txt
rq==0.12.0
redis==2.10.6
requests==2.20.0
pdfminer.six==20170720